got rid of the models file and moved the contents into the src/lib.rs file. Now only functionality related to the parsing and storing of data is in the lib file and the CLI related code is moved into main.rs
