+++ src/main.rs
@@ -2,15 +2,21 @@ extern crate csv_to;
 extern crate exitcode;
 extern crate structopt;
 
-use csv_to::CsvTo;
-use csv_to::cmd::db::{DbApp, Types};
-use csv_to::cmd::db::config::Config;
+//use csv_to::CsvTo;
+use csv_to::cmd::db;
+use csv_to::cmd::db::{
+    DbApp,
+    Types,
+    config::Config,
+};
 use csv_to::parsers::csv::CSVService;
 use csv_to::storage::{
     postgres::PostgresStore,
     sqlite::SQLiteStore
 };
 use structopt::StructOpt;
+use std::path::PathBuf;
+use csv_to::ParsedContent;
 
 fn main() {
     let opt = CsvTo::from_args();
@@ -57,4 +63,42 @@ fn main() {
             }
         }
     };
+}
+
+/// All command line options/flags broken into their sub-commands
+#[derive(Debug, StructOpt)]
+#[structopt(name = "csv-to", about = "creates databases and code from CSV data")]
+pub enum CsvTo {
+    #[structopt(name = "db", about = "creates and loads a database from CSV file(s)")]
+    Db {
+        #[structopt(short = "e", long = "extension", help = "the file extension for the CSV files to be parsed", default_value = "csv")]
+        extension: String,
+
+        #[structopt(short = "f", parse(from_os_str), long = "files", help = "The CSV files to be processed, can be /path/to/files/ or a comma delimited string of paths")]
+        files: Vec<PathBuf>,
+
+        #[structopt(short = "d", parse(from_os_str), long = "directories", help = "The directories that contain CSV files to be processed, a comma delimited string of paths")]
+        directories: Vec<PathBuf>,
+
+        #[structopt(short = "t", long = "type", help = "The type of database to create, valid types are sqlite and postgres")]
+        db_type: db::Types,
+
+        #[structopt(short = "c", long = "connection-info", help = "Database connectivity information")]
+        connection_info: String,
+
+        #[structopt(short = "n", long = "name", help = "Name of the database to be created")]
+        name: String,
+
+        #[structopt(long = "drop-stores", help = "Drops tables/collections if the already exist")]
+        drop_stores: bool,
+
+        #[structopt(long = "no-headers", help = "The CSV file(s) have no column headers")]
+        no_headers: bool
+    }
+}
+
+// This trait is what all of the sub-commands will implement so they can have a common
+// interface that the main can call into to start the csv_to logic started
+trait App {
+    fn run(&self) -> Result<ParsedContent, std::io::Error> ;
 }
